fix mirroring of base and bnb
